+++ src/components/common/AmountButtons/index.module.css
@@ -1,0 +1,20 @@
+.toBasket {
+  padding: 10px;
+  display: flex;
+}  
+
+.changeAmount, .addToBasket {
+  margin: 0 5px;
+  text-align: center;
+  line-height: 26px;
+  border: none;
+  cursor: pointer;
+}
+.changeAmount {
+  background-color: transparent;
+  /* width: 26px;
+  height: 26px; */
+  font-weight: 900;
+  font-size: 20px;
+
+}
