+++ src/components/common/Basket/Basket.module.css
@@ -1,13 +1,20 @@
+.basketIcon{
+  cursor: pointer;
+}
+
 .basket{
   position: absolute;
-  right: 5px;
-  width: 300px;
+  right: 25vw;
+  top: 80px;
+  width: 50vw;
+  
   height: 600px;
   background-color: white;
   outline: 1px solid blue;
 }
 
 .closeBasket {
+  cursor: pointer;
   background-color: gray;
   position: absolute;
   width: 30px;
