+++ src/components/common/BasketCard/index.module.css
@@ -1,0 +1,37 @@
+.root {
+  width: 90%;
+  outline: 1px solid pink;
+  height: 100px;
+  margin: 20px auto;
+
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+}
+
+.number{
+  font-size: 60px;
+}
+
+.description {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+}
+
+.title{
+  font-weight: bold;
+  font-size: 20px;
+  text-transform: capitalize;
+}
+
+.input {
+  font-size: 10px;
+  height: 50px;
+  resize: none;
+}
+
+.inputLabel {
+  font-size: 10px;
+  color: gray;
+}
